requirements.txt
redis==3.2.0
six==1.11.0
numpy==1.16.4
scipy==1.2.1
web.py==0.40
scipy==1.2.1
torchvision==0.5.0
opencv_python==3.4.5.20
opencv-contrib-python==3.4.2.17
Pillow==6.2.1
web.py==0.40
flask==1.0.2